+++ test/helper/supports.js
@@ -19,7 +19,7 @@ define(function(require) {
     canFocusScrollContainer: require('ally/supports/focus-scroll-container'),
     canFocusScrollContainerWithoutOverflow: require('ally/supports/focus-scroll-container-without-overflow'),
     canFocusSvg: require('ally/supports/focus-Svg'),
-    canFocusSvgMethod: Boolean(Element.prototype.focus),
+    canFocusSvgMethod: require('ally/supports/svg-focus-method'),
     canFocusoutEvent: require('ally/supports/focusout-event'),
     cssShadowPiercingDeepCombinator: require('ally/supports/css-shadow-piercing-deep-combinator'),
   };